Refactor back + MatrixView base
